many edits
style fixed, move to distance_measures, fix example
