--- doc/reference/algorithms/kemeny.rst
@@ -1,11 +1,0 @@
-*****************
-Distance Measures
-*****************
-
-.. automodule:: networkx.algorithms.kemeny
-.. autosummary::
-   :toctree: generated/
-
-   kemeny_constant
-
-
