+++ networkx/algorithms/__init__.py
@@ -29,7 +29,6 @@ from networkx.algorithms.link_analysis import *
 from networkx.algorithms.link_prediction import *
 from networkx.algorithms.lowest_common_ancestors import *
 from networkx.algorithms.isolate import *
-from networkx.algorithms.kemeny import *
 from networkx.algorithms.matching import *
 from networkx.algorithms.minors import *
 from networkx.algorithms.mis import *
